rewrite: moved tests to common.r4
