+++ src/Hl7.FhirPath.R4.Tests/Hl7.FhirPath.R4.Tests.csproj
@@ -17,6 +17,11 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
+
+  <ItemGroup>
+    <None Remove="TestData\bundle-contained-references.xml" />
+    <None Remove="TestData\Large-QuestionnaireResponse.xml" />
+  </ItemGroup>
   
   <ItemGroup>
     <ProjectReference Include="..\Hl7.Fhir.Core\Hl7.Fhir.Core.csproj" />
