+++ src/Hl7.FhirPath.R4.Tests/PocoTests/FhirPathExtensionsTest.cs
@@ -6,8 +6,6 @@
  * available at https://raw.githubusercontent.com/FirelyTeam/fhir-net-api/master/LICENSE
  */
 
-using System.Diagnostics;
-using Hl7.Fhir.Utility;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Hl7.Fhir.Model;
 using Hl7.Fhir.ElementModel;
