+++ src/Hl7.FhirPath.R4.Tests/TestData/bundle-contained-references.xml
@@ -1,0 +1,52 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<Bundle xmlns="http://hl7.org/fhir">
+  <type value="collection"/>
+
+  <!-- An organization that already has an id on a server -->
+  <entry>
+    <fullUrl value="http://example.org/fhir/Organization/23"/>
+    <resource>
+      <Organization>
+        <id value="23"/>
+        <active value="true" />
+      </Organization>
+    </resource>
+  </entry>
+
+  <!-- An organization that doesn't have a persistent home - but it does have 
+      a UUID assigned for this bundle "locally identified" -->
+  <entry>
+    <fullUrl value="urn:uuid:04121321-4af5-424c-a0e1-ed3aab1c349d"/>
+    <resource>
+      <Organization>
+        <active value="true" />
+      </Organization>
+    </resource>
+  </entry>
+
+
+  <!-- a contained reference -->
+  <entry>
+    <fullUrl value="http://example.org/fhir/Patient/e"/>
+    <resource>
+      <Patient>
+        <id value="e"/>
+        <contained>
+          <Organization>
+            <id value="orgX" />
+          </Organization>
+        </contained>
+        <contained>
+          <Organization>
+            <id value="orgY" />
+          </Organization>
+        </contained>
+        <managingOrganization>
+          <!-- reference to the contained organization -->
+          <reference value="#orgY"/>
+        </managingOrganization>
+      </Patient>
+    </resource>
+  </entry>
+
+</Bundle>
